app/auth/auth.py
from passlib.context import CryptContext
from datetime import datetime, timedelta
from jose import jwt 


SECRET_KEY = "Some_key_which_i_dont_know"
ALGORITHM = "HS256"
ACCESS_TOEKN_EXPIRE_MINUTE = 60

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


def hash_password(password: str):
    return pwd_context.hash(password)


def verify_password(plain_password: str, hashed_password: str):
    return pwd_context.verify(plain_password, hashed_password)


def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOEKN_EXPIRE_MINUTE)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

    return encoded_jwt